Types: Split `SymbolTable` and `Scope` into separate source files
