+++ loki/types/__init__.py
@@ -11,3 +11,4 @@ The Loki type system to store and shared type information on symbols and IR node
 
 from loki.types.datatypes import *  # noqa
 from loki.types.scope import *  # noqa
+from loki.types.symbol_table import *  # noqa
